projects/Document_analyse/src/requirements.txt
azure.core
azure.storage.blob
azure-ai-documentintelligence
python-dotenv
streamlit